Backend: Add daily scheduling of value writing
Events are constantly being detected over the serial connection.
These events get saved with a timestamp.
After a day, a Dataframe with all the events of that day will be created.
The internal state of the core will be reset, and writing for the next day
will start. A writing period will be started at midnight and lasts one
whole day.