Backend: Add daily and weekly csv write mechanism
1. The Update for daily and weekly values will be scheduled every midnight.
An average value of the number of customers per time interval will
be saved to a file, including the number of passed days.
These values will be used in combination with the values of the current
day to calculate new average daily values.
The maximum number of customers is also saved to a file.

2. The average values per day and the maximum values of the current week
can be requested from the REST API.

3. File reading and writing now uses pandas Dataframes to allow for easier
conversions and filtering of values.

4. Additional requirements for libraries has been added